Removed pause button, allowed aborting
Removed pause button.
Quietened occasional beeping when keyboard buttons are pressed while the
focus is on an inactive item of the window.
Allowed discard record button to be active during recording, to serve as
an abort button.
Removed a lot of junk files.
(Possibly added a lot of junk files too - hard to tell with Windows...)